separando classes alterando retorno para completion
