+++ Data/Http/HttpPostClient.swift
@@ -9,5 +9,5 @@ import Foundation
 import Domain
 
 public protocol HttpPostClient {
-    func post(to url: URL, with data: Data?, completion: @escaping (HttpError) -> Void)
+    func post(to url: URL, with data: Data?, completion: @escaping (Result<Data, HttpError>) -> Void)
 }
