+++ Data/UseCases/RemoteAddAccount.swift
@@ -8,7 +8,7 @@
 import Domain
 import Foundation
 
-public final class ApiAddAccount {
+public final class RemoteAddAccount: AddAccount {
     private let url: URL
     private let httpClient: HttpPostClient
 
@@ -17,9 +17,9 @@ public final class ApiAddAccount {
         self.httpClient = httpClient
     }
     
-    public func add(addAccountModel: AddAccountModel, completion: @escaping (DomainError) -> Void) {
+    public func add(addAccountModel: AddAccountModel, completion: @escaping (Result<AccountModel, DomainError>) -> Void) {
         httpClient.post(to: url, with: addAccountModel.toData()) { error in
-            completion(.unexpected)
+            completion(.failure(.unexpected))
         }
     }
 }
