+++ DataTests/Services/HttpClientSpy.swift
@@ -1,0 +1,25 @@
+//
+//  HttpClientSpy.swift
+//  DataTests
+//
+//  Created by Denis Couras on 01/12/21.
+//
+
+import Data
+import Domain
+import Foundation
+
+class HttpClientSpy: HttpPostClient {
+    var urls = [URL]()
+    var data: Data?
+    var completion: ((Result<Data, HttpError>) -> Void)?
+    func post(to url: URL, with data: Data?, completion: @escaping (Result<Data, HttpError>) -> Void) {
+        self.urls.append(url)
+        self.data = data
+        self.completion = completion
+    }
+    
+    func completeWithError(_ error: HttpError) {
+        completion?(.failure(error))
+    }
+}
