+++ DataTests/UseCases/ApiAddAccountTests.swift
@@ -25,8 +25,13 @@ class ApiAddAccountTests: XCTestCase {
     func testAddShouldCompleteWithErrorIfClientFails() {
         let (sut, httpClient) = ApiAddAccountMock().makeSut()
         let exp = expectation(description: "waiting")
-        sut.add(addAccountModel: AddAccountModelMock().makeAddAccountModel()) { error in
-            XCTAssertEqual(error, .unexpected)
+        sut.add(addAccountModel: AddAccountModelMock().makeAddAccountModel()) { result in
+            switch result {
+            case .failure(let error):
+                XCTAssertEqual(error, .unexpected)
+            case .success:
+                XCTFail("Expected error receive \(result) instead")
+            }
             exp.fulfill()
         }
         httpClient.completeWithError(.noConnection)
@@ -34,35 +39,4 @@ class ApiAddAccountTests: XCTestCase {
     }
 }
 
-class HttpClientSpy: HttpPostClient {
-    var urls = [URL]()
-    var data: Data?
-    var completion: ((HttpError) -> Void)?
-    func post(to url: URL, with data: Data?, completion: @escaping (HttpError) -> Void) {
-        self.urls.append(url)
-        self.data = data
-        self.completion = completion
-    }
-    
-    func completeWithError(_ error: HttpError) {
-        completion?(error)
-    }
-}
 
-class ApiAddAccountMock {
-    func makeSut(url: URL = URL(string: "http://any.com")!) -> (sut: ApiAddAccount, httpClientSpy: HttpClientSpy) {
-        let httpClient = HttpClientSpy()
-        return (ApiAddAccount(url: url, httpClient: httpClient), httpClient)
-    }
-}
-
-class AddAccountModelMock {
-    func makeAddAccountModel() -> AddAccountModel {
-        AddAccountModel(
-            name: "any_name",
-            email: "any@email.com",
-            password: "pass",
-            passwordConfirmation: "pass"
-        )
-    }
-}
